subscriber/requirements.txt
python-dotenv
SQLAlchemy
pyyaml
pytz
PyMySQL==1.0.2
easydict
pynostr[websocket-client]
flask